+++ src/control_plane/migration/object_store/mod.rs
@@ -3,6 +3,10 @@ mod tests;
 
 mod backup_minio_bucket;
 mod minio_backup_options;
+mod minio_restore_options;
+mod restore_minio_bucket;
 
 pub(crate) use backup_minio_bucket::backup_minio_bucket;
 pub(crate) use minio_backup_options::MinioBackupOptions;
+pub(crate) use minio_restore_options::MinioRestoreOptions;
+pub(crate) use restore_minio_bucket::restore_minio_bucket;